+++ Contrib/AnyUnit.Style.Nunit/TestFixtureAttribute.cs
@@ -20,6 +20,7 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using AnyUnit.Util;
+using AnyUnit.Run.Attributes;
 
 namespace AnyUnit.Style.Nunit
 {